Fix namespaces for local dead store elimination
Declare LocalDeadStoreElimination class to be under TR namespace,
eliminating the need to map it from OMR to TR.
